requirements.txt
Flask==3.0.3
Pillow==10.4.0
flask-caching==2.3.0
python-dotenv==1.0.1
flask-cors==5.0.10
gunicorn==23.0.0